pyproject.toml
[project]
name = "rag-project"
version = "0.1.0"
requires-python = ">=3.10, <3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "chardet>=5.2.0",
    "chromadb>=1.0.12",
    "dotenv>=0.9.9",
    "ebooklib>=0.19",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "google-genai>=1.20.0",
    "llama-index>=0.12.42",
    "llama-index-embeddings-openai-like>=0.1.1",
    "llama-index-llms-gemini>=0.5.0",
    "llama-index-llms-google-genai>=0.2.1",
    "llama-index-llms-openai-like>=0.4.0",
    "llama-index-readers-web>=0.4.2",
    "llama-index-vector-stores-chroma>=0.4.2",
    "lxml>=5.4.0",
    "markdown>=3.8",
    "openai>=1.88.0",
    "openpyxl>=3.1.5",
    "pdfplumber>=0.11.7",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
    "rag-engine",
    "setuptools>=80.9.0",
    "xlrd>=2.0.2",
]

[tool.uv.sources]
rag-engine = { path = "rag_engine", editable = true }
